refactor(cli): extract connectivity probes from doctor into doctor-probes
